% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ssm.R
\name{buildDLM}
\alias{buildDLM}
\title{Estimate values of \eqn{k_{1,t}} in \eqn{F_t} of the DLM and calculate the
covariance matrices to be used when have to calculate the transition
probabilities.}
\usage{
buildDLM(iniDLM, Y, k1 = 0.044)
}
\arguments{
\item{iniDLM}{The DLM created using \link{iniDLM}.}

\item{Y}{A matrix with 2 observation rows (weight and intake). One column for each week. Used to estimate
\eqn{k_{1,t}}. Could be mean values observed in the herd.}

\item{k1}{Value of \eqn{k_1} used in the estimation.}
}
\value{
A list containing all the parameters. Element L contains the covariance matrix for transition probabilities.
  Element JFF is a (2x2) matrix indicating where \eqn{F_t'} depends on \eqn{t} (i.e. a 1 is put where \eqn{k_t}) is
  included). Element X contains the \eqn{k_t} values. See the vignette (\code{vignette("dlm")}) in the \pkg{dlm}
  package for further info about the specification.
}
\description{
Estimate using formula \eqn{k_{1,t} = k_1*a_t^(-0.25)} where \eqn{a_t} is the prior at time \eqn{t}.
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk} and Lars Relund \email{lars@relund.dk}
}

