% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{calcFullObs}
\alias{calcFullObs}
\title{Update the sample data with "pseudo" observations (observations if all pigs in the pen
were present) based on the formulars of the truncted normal distribution
(Herd management book, Vol 1, p 128)}
\usage{
calcFullObs(dat)
}
\arguments{
\item{dat}{Weekly sample data table.}
}
\value{
Updated data table with additional observation columns (ending with Calc).
}
\description{
Update the sample data with "pseudo" observations (observations if all pigs in the pen
were present) based on the formulars of the truncted normal distribution
(Herd management book, Vol 1, p 128)
}
\author{
Lars Relund \email{lars@relund.dk}
}

