% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{calcWeekSampleData}
\alias{calcWeekSampleData}
\title{Calc a weekly data set based on the daily samples. We assume that actions are
considered on Mondays (in the morning), i.e week w in the HMDP corresponds week w-1 and
weight data to data from sunday the week before!}
\usage{
calcWeekSampleData(dtPigs, measurementsPerDay, measurementStd)
}
\arguments{
\item{measurementsPerDay}{Measurements per day from the weigthing sensor (normally vision).}

\item{measurementStd}{The std of the weight sensor (normally vision).}
}
\value{
A data table with a row for each week with columns \code{aveOLW<X>} (ave observed
  liveweight at t), \code{sdOLW<X>} (sd observed live weight at t), \code{aveFI<X>} (ave feed
  intake per pig per week, monday-sunday, we consider sundays)

the number of pigs, the average feed intake,
  ave weight and sd at the given day in the week. Both data for the selected and uselected pen
  are given. Note the observations of ave weight and sd are based on pigs randomly selected and
  mesured (often more than once) and hence may vary. The true weight and sd are based on the true
  weights of each pig.
}
\description{
Calc a weekly data set based on the daily samples. We assume that actions are
considered on Mondays (in the morning), i.e week w in the HMDP corresponds week w-1 and
weight data to data from sunday the week before!
}
\examples{
dat<-calcWeekSampleData()
}
\author{
Lars Relund \email{lars@relund.dk}
}

