% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{changeFeedMixAction}
\alias{changeFeedMixAction}
\title{Resimulate based on a new feed-mix action at a given stage.}
\usage{
changeFeedMixAction(param, dTList, stage, feedMix)
}
\arguments{
\item{dTList}{A list returned by \link{simulatePen}}

\item{stage}{Stage number.}

\item{feedMix}{New feed-mix.}
}
\value{
The list dTList modified so pigs have been culled and weekly data recalculated.
}
\description{
Resimulate based on a new feed-mix action at a given stage.
}
\examples{
dat<-thresholdAction(dat,5,90)
}
\author{
Lars Relund \email{lars@relund.dk}
}

