% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{createDailySampleData}
\alias{createDailySampleData}
\title{Create a data set simulating a pen from DFI and T days ahead.}
\usage{
createDailySampleData(pen, pigIds, TLW, culled, measurementStd, k4Mean,
  DGFactor, T, DFI, feedMix)
}
\arguments{
\item{pen}{Pen id added to the output.}

\item{pigIds}{Id of the pigs simulated}

\item{TLW}{Start weight day before simulate (same length as pigIds).}

\item{culled}{Have the pig we simulate been culled (same length as pigIds).}

\item{measurementStd}{The std of the weight sensor (normally vision).}

\item{k4Mean}{Mean of growth rate parameter (Gompertz growth curve).}

\item{DGFactor}{Increase the daily gain (x2) with this factor.}

\item{T}{Number of days to simulate.}

\item{DFI}{Days from insertion into the pen (start simulation at t=DFI+1).}

\item{feedMix}{Feed-mix number used.}
}
\value{
A data table containing a row for each pig on a daily basis with columns \code{t} (days
  since insertion into the pen), \code{pig} (pig id), \code{FI} (feed intake in feed units),
  \code{DG} (daily gain), \code{TLW} (true live weight), \code{OLW} (observed live weight),
  \code{SW} (slaughter weight), \code{culled} (are we simulating a culled pig (=0, alive=1),
  \code{week} (week number), \code{weekDay} (week day - monday = 1), \code{pen} (pen id),
  \code{feedMix} (feed-mix number used).
}
\description{
Create a data set simulating a pen from DFI and T days ahead.
}
\examples{
dat<-createDailySampleData()
}
\author{
Lars Relund \email{lars@relund.dk}
}

