% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{DGLMfilter}
\alias{DGLMfilter}
\title{nGSSM filtering to estimate the weight variances during the growing period.}
\usage{
DGLMfilter(mod1, ob, W)
}
\arguments{
\item{mod1}{set of parameters of the nGSSM}

\item{ob}{set of observations related to the sample variances of the weights during the growing period.}

\item{W}{System variance of the nGSSM}
}
\value{
Updated information of posterior.
}
\description{
nGSSM filtering to estimate the weight variances during the growing period.
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk}
}

