% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{DLMfilter}
\alias{DLMfilter}
\title{GSSM filtering to estimate the weight and growth during the growing period.}
\usage{
DLMfilter(mod, D, W)
}
\arguments{
\item{mod}{set of parameters needed in the GSSM model.}

\item{D}{Set of the weight and feed intake data during the growing period.}

\item{W}{system variance of the GSSM.}
}
\value{
Updated information of posterior.
}
\description{
GSSM filtering to estimate the weight and growth during the growing period.
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk}
}

