% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{estimateK4}
\alias{estimateK4}
\title{Fit parameter k4 as a function of average daily growth (over a period of 84 days).
Used to find different k4's for different feed-mix.
Based on the appendix of Jørgensen (1993).}
\usage{
estimateK4(ite = 500)
}
\arguments{
\item{ite}{Number of iterations (pigs used in the estimation per k4 value).}
}
\value{
A data frame with k4 values given a set of daily gains
}
\description{
Fit parameter k4 as a function of average daily growth (over a period of 84 days).
Used to find different k4's for different feed-mix.
Based on the appendix of Jørgensen (1993).
}
\examples{
dat<-estimateK4()
plot(dat)
}
\author{
Lars Relund \email{lars@relund.dk}
}
\seealso{
samplePig.
}

