% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{findActions}
\alias{findActions}
\title{Add the optimal actions to a data set.}
\usage{
findActions(pen, paramSim, param, policy, startT, phase, feedMix, startFeed)
}
\arguments{
\item{policy}{Optimal policy acquired from the hmdp . We need to run HMDP and get the optimal policy}

\item{startT}{Starting time of using the DLM and the DGLM (starting time after having a new decision)}

\item{phase}{The phase number of feeding in the system at time "startT"}

\item{feedMix}{The feed-mix that is used in week "startT"}

\item{startFeed}{The time that the last feedMix has been started}

\item{dat}{Data table with estimated values (columns eAveOLW, eSdOLW and eAveG).}

\item{dat}{Data table with observed values (columns aveOLW, sdOLW and aveFI).}
}
\value{
An extended data table with new columns threshold (NA if not used) and feedMix (feed-mix used).
}
\description{
Add the optimal actions to a data set.
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk}
}

