% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setParam.R
\name{findInterval}
\alias{findInterval}
\title{Find the intervals the priorGrowth belongs to and return the index (c++ style)}
\usage{
findInterval(rations, priorGrowth, gIntervals)
}
\arguments{
\item{priorGrowth}{Prior growth estimates for each feed mix.}

\item{gIntervals}{Discretization list.}
}
\description{
Find the intervals the priorGrowth belongs to and return the index (c++ style)
}
\author{
Lars Relund \email{lars@relund.dk}
}

