% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{GrowthParam}
\alias{GrowthParam}
\title{Compute the growth rate parameter of the Gompertz function based on the current live weight and the daily gain.}
\usage{
GrowthParam(W, G, K)
}
\arguments{
\item{W}{weight at the current week in the production system (we suppose weight is the insertion weight in the system)}

\item{G}{Daily gain in kg}

\item{K}{Logarithm of the outgrowth weight in kg}
}
\value{
The growth rate parameter in the Gompertz function.
}
\description{
Compute the growth rate parameter of the Gompertz function based on the current live weight and the daily gain.
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk}
}

