% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iniDGLM}
\alias{iniDGLM}
\title{Initialize the parameters for the DGLM}
\usage{
iniDGLM(tMax = 15, nf = 35, W = 0, alpha0 = 130, beta0 = 17, c0 = 1)
}
\arguments{
\item{tMax}{Maximum growth period in weeks.}

\item{nf}{Sample size to estimate the sample variance in the pen (we suppose this value is constant and the sample is taken from the herd)}

\item{W}{System variance (we suppose this value is 0 in the DGLM)}

\item{alpha0}{Initial shape parameter at time t_0}

\item{beta0}{Initial rate parameter at time t_0}

\item{c0}{coefficient of G_t in the dglm model (Based on the simulation test, this value is 1)}
}
\value{
A list containing all the parameters
}
\description{
Initialize the parameters for the DGLM
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk} and Lars Relund \email{lars@relund.dk}
}

