% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{plotPigData}
\alias{plotPigData}
\title{Plot simulation data.}
\usage{
plotPigData(dat)
}
\arguments{
\item{dat}{Simulation data obtained from \link{samplePig}.}
}
\value{
NULL (a plot is output).
}
\description{
Plot simulation data.
}
\examples{
plotPigData(dat)
}
\author{
Lars Relund \email{lars@relund.dk}
}
\seealso{
createSampleData.
}

