% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pNorm1D}
\alias{pNorm1D}
\title{Univariate cumulative normal (distribution function)}
\usage{
pNorm1D(x, mean, sd)
}
\arguments{
\item{x}{Value of quantiles.}

\item{mean}{The mean.}

\item{sd}{The standard deviation.}
}
\value{
The distribution function.
}
\description{
Based on the paper "Better approximations to cumulative normal functions" by
G. West \url{http://www.wilmott.com/pdfs/090721_west.pdf}. The source code
has been taken from the authors webpage \url{http://finmod.co.za/research.html}.
}
\examples{
# test against pnorm
times<-500
x<-runif(times,-10,10)
mean<-0
sd<-3
diff<-abs(pnorm(x,mean,sd)-pNorm1D(x,mean,sd))
if (all(diff<1e-15)) cat("No problems")

# compare performance
testpnorm<-function() {
   x<-runif(1,-10,10)
   mean<-runif(1,-10,10)
   sd<-runif(1,0,5)
   pnorm(x,mean,sd)
}
testpNorm1D<-function() {
   x<-runif(1,-10,10)
   mean<-runif(1,-10,10)
   sd<-runif(1,0,5)
   pNorm1D(x,mean,sd)
}

require(rbenchmark)
benchmark(testpnorm(), testpNorm1D(), order="relative", replications=100000)[,1:4]
}
\author{
Lars Relund \email{lars@relund.dk}
}

