% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pNorm2D}
\alias{pNorm2D}
\title{Bivariate cumulative normal (distribution function)}
\usage{
pNorm2D(lower, upper, mean, sigma)
}
\arguments{
\item{lower}{The vector of lower limits (2-dim vector).}

\item{upper}{The vector of upper limits (2-dim vector).}

\item{mean}{The mean vector (2-dim vector).}

\item{sigma}{The covariance matrix (2x2).}
}
\value{
The distribution function.
}
\description{
Based on the paper "Better approximations to cumulative normal functions" by
G. West \url{http://www.wilmott.com/pdfs/090721_west.pdf}. The source code
has been taken from the authors webpage \url{http://finmod.co.za/research.html}.
}
\details{
Transform into a bivariate with sd = 1 using Y = 1/sqrt(diag(sigma))(X-mean), i.e.
Pr(X<=x) = P(Y<=1/sqrt(diag(sigma))(x-mean)).
}
\examples{
# test against pmvnorm
times<-500
diff<-rep(NA,times)
diff_arma<-rep(NA,times)
for (i in 1:times) {
   sigma <- matrix(c(runif(1,1,10),1,1,runif(1,1,10)),nrow=2)
   sigma[1,2]<-sigma[2,1]<-sqrt(sigma[1,1]*sigma[2,2])*runif(1,-1,1)
   mean <- c(runif(1,1,10),runif(1,1,10))
   lower<-c(runif(1,-6,6),runif(1,-6,6))
   upper<-c(runif(1,lower[1],6),runif(1,lower[2],6))
   p1<-pmvnorm(lower=lower,upper=upper,mean=mean,sigma=sigma)
   p2<-pNorm2D(lower, upper, mean, sigma)
   p3<-pNorm2D_arma(lower, upper, mean, sigma)
   diff[i]<-abs(p1-p2)
   diff_arma[i]<-abs(p1-p3)
}
if (all(diff<1e-15)) cat("No problems with pNorm2D") else stop("Results for pNorm2D not the same!")
if (all(diff_arma<1e-15)) cat("No problems with pNorm2D_arma") else stop("Results for pNorm2D_arma not the same!")

# compare performance
testPmvnorm<-function() {
   sigma <- matrix(c(runif(1,1,10),1,1,runif(1,1,10)),nrow=2)
   sigma[1,2]<-sigma[2,1]<-sqrt(sigma[1,1]*sigma[2,2])*runif(1,-1,1)
   mean <- c(runif(1,1,10),runif(1,1,10))
   lower<-c(runif(1,-6,6),runif(1,-6,6))
   upper<-c(runif(1,lower[1],6),runif(1,lower[2],6))
   pmvnorm(lower=lower,upper=upper,mean=mean,sigma=sigma)
}
testpNorm2D<-function() {
   sigma <- matrix(c(runif(1,1,10),1,1,runif(1,1,10)),nrow=2)
   sigma[1,2]<-sigma[2,1]<-sqrt(sigma[1,1]*sigma[2,2])*runif(1,-1,1)
   mean <- c(runif(1,1,10),runif(1,1,10))
   lower<-c(runif(1,-6,6),runif(1,-6,6))
   upper<-c(runif(1,lower[1],6),runif(1,lower[2],6))
   pNorm2D_arma(lower, upper, mean, sigma)
}
testpNorm2D_arma<-function() {
   sigma <- matrix(c(runif(1,1,10),1,1,runif(1,1,10)),nrow=2)
   sigma[1,2]<-sigma[2,1]<-sqrt(sigma[1,1]*sigma[2,2])*runif(1,-1,1)
   mean <- c(runif(1,1,10),runif(1,1,10))
   lower<-c(runif(1,-6,6),runif(1,-6,6))
   upper<-c(runif(1,lower[1],6),runif(1,lower[2],6))
   pNorm2D_arma(lower, upper, mean, sigma)
}

require(rbenchmark)
benchmark(testPmvnorm(), testpNorm2D(), testpNorm2D_arma, order="relative", replications=10000)[,1:4]
}
\author{
Lars Relund \email{lars@relund.dk}
}

