% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setParam.R
\name{PointsW}
\alias{PointsW}
\title{Descritize the state space of the HMDP based on the number of the stage: Subfunction : Find the centerpoints for the live weight
and stansard deviation}
\usage{
PointsW(week, numPoints, length, sdWeights, meanWeights)
}
\arguments{
\item{length}{The length of the related intervals in the current styage}

\item{Week}{The current stage (week) in the production system}
}
\value{
A vector included the center points to discritize the state space in the current stage
}
\description{
Descritize the state space of the HMDP based on the number of the stage: Subfunction : Find the centerpoints for the live weight
and stansard deviation
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk}
}

