% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{samplePig}
\alias{samplePig}
\title{Create a simulation of one pig on a daily basis.
The function uses the model described in the appendix of Jørgensen (1993)
'The Influence of Weighing Precision on Delivery Decisions in Slaughter Pig Production'}
\usage{
samplePig(pig = 1, measurementStd = 2, T = 7 * 11, k4Mean = 0.0166,
  TLW = rnorm(1, 30, measurementStd), DFI = 0, DGFactor = 1, culled = 0)
}
\arguments{
\item{pig}{The id to be used of the pig (integer). Usefull when merge multiple pigs into one data table.}

\item{measurementStd}{The std of the weight sensor (e.g vision).}

\item{T}{Last day to simulate}

\item{k4Mean}{Mean of growth rate parameter (Gompertz growth curve).}

\item{TLW}{Simulation start weight (day before start simulation t=0).}

\item{DFI}{Days from insertion into the pen (start simulation at t=DFI+1).}

\item{DGFactor}{Increase the daily gain (x2) with this factor.
Can be used to indicate genetics of higher growth.}

\item{culled}{Do we simulate a culled pig (true(1)/false(0)).}
}
\value{
A \link(data.table) with columns \code{t} (days since insertion into the pen),
\code{pig} (pig id), \code{FI} (feed intake in feed units), \code{DG} (daily gain),
\code{TLW} (true live weight),
\code{OLW} (observed live weight), \code{SW} (slaughter weight),
\code{culled} (are we simulating a culled pig (=0, alive=1).
}
\description{
Create a simulation of one pig on a daily basis.
The function uses the model described in the appendix of Jørgensen (1993)
'The Influence of Weighing Precision on Delivery Decisions in Slaughter Pig Production'
}
\examples{
dat<-samplePig()
plotPigData(dat)
}
\author{
Lars Relund \email{lars@relund.dk}
}
\seealso{
createDailySampleData.
}

