% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{setGSSM}
\alias{setGSSM}
\title{Set the parameters of GSSM.}
\usage{
setGSSM(t = 12, FF, GG, V, W, m0, C0, D)
}
\arguments{
\item{t}{Maximumu life time of the pen}

\item{FF}{design matrix of system equation.}

\item{GG}{design matrix of observation equation.}

\item{V}{Observation variance of the nGSSM}

\item{W}{System variance of the nGSSM}

\item{m0}{Initial mean of posterior distribution at the insertion time (t=0)}

\item{C0}{Initial variance of posterior distribution at the insertion time (t=0)}
}
\value{
A list containing the parameters of the GSSM.
}
\description{
Set the parameters of GSSM.
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk}
}

