% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nGSSM.R
\name{setnGSSM}
\alias{setnGSSM}
\title{Set the parameters of nGSSM}
\usage{
setnGSSM(t = 12, nf, W, c0, d0)
}
\arguments{
\item{t}{Maximumu life time of the pen}

\item{nf}{Number of weight observations}

\item{W}{System variance of the nGSSM}

\item{c0}{Initial shape parameter of the prior distribution (Inv-Gamma distribution)}

\item{d0}{Initial scale parameter of the prior distribution (Inv-Gamma distribution)}
}
\value{
A list containing the recquired parameters.
}
\description{
Set the parameters of nGSSM
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk}
}

