% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{setSimParam}
\alias{setSimParam}
\title{Set the simulation parameters for a pen.}
\usage{
setSimParam(pen = 1, pigIds = 1:15, measurementsPerDay = 30,
  measurementStd = 1, k4Mean = c(0.01012228, 0.01150325, 0.01387502,
  0.01487517), DGFactor = 1)
}
\arguments{
\item{pen}{Pen id added to the output.}

\item{pigIds}{Id of the pigs simulated}

\item{measurementsPerDay}{Measurements per day from the weigthing sensor (normally vision).}

\item{measurementStd}{The std of the weight sensor (normally vision).}

\item{k4Mean}{Mean of growth rate parameter (Gompertz growth curve). One value for each possible feed-mix.}

\item{DGFactor}{Increase the daily gain (x2) with this factor.}
}
\value{
A list containing the parameters.
}
\description{
Set the simulation parameters for a pen.
}
\examples{
dat<-simulatePen()
}
\author{
Lars Relund \email{lars@relund.dk}
}

