% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulatePen}
\alias{simulatePen}
\title{Create a data set simulating a pen from DFI and T days ahead. We assume that actions are
considered on Mondays (in the morning), i.e week w in the HMDP corresponds week w-1 and
weight data to data from sunday the week before!}
\usage{
simulatePen(param, feedMix, TLW = rnorm(length(param$pigIds), 30,
  param$measurementStd), culled = rep(0, length(param$pigIds)), T = 7 * 11,
  DFI = 0)
}
\arguments{
\item{param}{Pen simulation parameters.}

\item{feedMix}{Feed-mix used.}

\item{TLW}{Start weight day before simulate (same length as pigIds).}

\item{culled}{Have the pig we simulate been culled (same length as pigIds).}

\item{T}{Number of days to simulate.}

\item{DFI}{Days from insertion into the pen (start simulation at t=DFI+1)}
}
\value{
A list with two data tables. Table \code{dtDailyPig} contains a row
for each pig on a daily basis. Table \code{dtWeekAve}
contain a row for each week with the numbe of pigs, the average feed
intake, ave weight and sd at the given day in the week.
}
\description{
Create a data set simulating a pen from DFI and T days ahead. We assume that actions are
considered on Mondays (in the morning), i.e week w in the HMDP corresponds week w-1 and
weight data to data from sunday the week before!
}
\examples{
dat<-simulatePen()
}
\author{
Lars Relund \email{lars@relund.dk}
}

