% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulation.R
\name{thresholdAction}
\alias{thresholdAction}
\title{Cull pigs based on a threshold action at a given stage.}
\usage{
thresholdAction(param, dTList, stage, th)
}
\arguments{
\item{dTList}{A list similar to the one returned by \link{simulatePen}}

\item{stage}{Stage number.}

\item{th}{Threshold.}
}
\value{
The list dTList modified so pigs have been culled and weekly data recalculated.
}
\description{
Cull pigs based on a threshold action at a given stage.
}
\examples{
dat<-thresholdAction(dat,5,90)
}
\author{
Lars Relund \email{lars@relund.dk}
}

