% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ssm.R
\name{iniDGLM}
\alias{iniDGLM}
\title{Initialize the parameters for the DGLM}
\usage{
iniDGLM(tMax = 15, nf = 35, W = 0, s0 = 7.65)
}
\arguments{
\item{tMax}{Maximum growth period in weeks.}

\item{nf}{Sample size to estimate the sample variance in the pen (we suppose this value is constant and the sample is taken from the herd)}

\item{W}{System variance (we suppose this value is 0 in the DGLM)}

\item{s0}{Initial sample variance}
}
\value{
A list containing all the parameters
}
\description{
Initialize the parameters for the DGLM
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk} and Lars Relund \email{lars@relund.dk}
}

